+++ ext/ClimaCalibrateClimaAnalysisExt.jl
@@ -11,7 +11,7 @@ import Statistics
 import Statistics: mean
 import NaNStatistics: nanmean, nanvar
 
-import LinearAlgebra: Diagonal, I, diagind
+import LinearAlgebra: Diagonal, I, diag, diagind
 
 # Used by both by SampleBuilder and ObservationRecipe for default flattening
 # of OutputVars