+++ test/runtests.jl
@@ -11,6 +11,7 @@ using SafeTestsets
 @safetestset "HPC job scripts" begin include("hpc_job_scripts.jl") end
 @safetestset "Workers per node" begin include("workers_per_node.jl") end
 @safetestset "Sampler" begin include("sample_builder.jl") end
+@safetestset "Diagonal term" begin include("diagonal_term.jl") end
 @safetestset "Observation recipe" begin include("observation_recipe.jl") end
 @safetestset "Ensemble builder" begin include("ensemble_builder.jl") end
 @safetestset "SVD analysis" begin include("svd_analysis.jl") end